Handle RestClient::ExceptionWithResponse without http_code
There seem to be cases where this class of errors are thrown without an
HTTP Code. We have experienced this on our live deployments, but were
unable to pinpoint the actual error.

Instead we get

```
undefined method `>' for nil:NilClass
```

By code provided on [rest-client](https://github.com/rest-client/rest-client/tree/v2.0.2#response-callbacks-error-handling) it seems we should handle this.